requirements.txt
requests
loguru
vk_api
python_dotenv
